Refactor ID generation: implement SequentialIdGenerator and update processor logic
