+++ tests/test_config_cliapp.py
@@ -24,6 +24,7 @@ def test_cli_app_config_process(cli_app_config):
         "ProjectX", # name
         "tST",      # prefix
         "pyt",      # scope
+        "random",   # numbering_mode
     ])
 
     def mock_input(prompt):
